+++ modules/planning/toolkits/optimizers/road_graph/dp_road_graph.h
@@ -93,8 +93,8 @@ class DpRoadGraph {
     common::SLPoint sl_point;
     const DpRoadGraphNode *min_cost_prev_node = nullptr;
     ComparableCost min_cost = {true, true, true,
-                               std::numeric_limits<float>::infinity(),
-                               std::numeric_limits<float>::infinity()};
+                               std::numeric_limits<double>::infinity(),
+                               std::numeric_limits<double>::infinity()};
     QuinticPolynomialCurve1d min_cost_curve;
   };
   bool GenerateMinCostPath(const std::vector<const Obstacle *> &obstacles,
@@ -103,8 +103,8 @@ class DpRoadGraph {
   bool IsValidCurve(const QuinticPolynomialCurve1d &curve) const;
 
   void GetCurveCost(TrajectoryCost trajectory_cost,
-                    const QuinticPolynomialCurve1d &curve, const float start_s,
-                    const float end_s, const uint32_t curr_level,
+                    const QuinticPolynomialCurve1d &curve, const double start_s,
+                    const double end_s, const uint32_t curr_level,
                     const uint32_t total_level, ComparableCost *cost);
 
   struct RoadGraphMessage {
